requirements.txt
asn1crypto                    == 0.24.0
bson                          == 0.5.8
certifi                       == 2019.3.9
cffi                          == 1.12.2
chardet                       == 3.0.4
coreapi                      ==  2.3.3
coreschema                    == 0.0.4
cryptography                  == 2.6.1
decorator                     == 4.0.11
Django                        == 2.0.13
django-ckeditor               == 5.7.1
django-cors-headers           == 2.5.2
django-environ                == 0.4.5
django-filter                 == 2.1.0
django-guardian               == 1.5.0
django-js-asset               == 1.2.2
django-rest-auth              == 0.9.5
django-rest-swagger           == 2.2.0
django-simple-history         == 2.7.3
django-telegrambot            == 1.0.1
djangorestframework           == 3.9.2
djangorestframework-guardian  == 0.1.1
djangorestframework-jwt       == 1.11.0
djangorestframework-simplejwt == 4.1.0
env                           == 0.1.0
environ                       == 1.0
envparse                      == 0.2.0
future                        == 0.17.1
get                           == 2019.3.22
idna                          == 2.8
imageio                       == 2.1.2
InstagramAPI                  == 1.0.2
itypes                        == 1.1.0
Jinja2                        == 2.10.1
jsonfield                     == 2.0.2
MarkupSafe                    == 1.1.1
moviepy                       == 0.2.3.2
numpy                         == 1.16.3
openapi-codec                 == 1.3.2
opencv-python                 == 4.1.0.25
passlib                       == 1.7.1
Pillow                        == 5.4.1
pip                           == 19.2.1
pkg-resources                 == 0.0.0
post                          == 2019.3.22
public                        == 2019.3.22
pycparser                     == 2.19
PyJWT                         == 1.7.1
PyMySQL                       == 0.9.3
pytesseract                   == 0.2.7
python-dateutil               == 2.8.0
python-telegram-bot           == 11.1.0
pytz                          == 2018.9
query-string                  == 2019.3.22
request                       == 2019.3.22
requests                      == 2.11.1
requests-toolbelt             == 0.7.0
setuptools                    == 39.0.1
simplejson                    == 3.16.0
six                           == 1.12.0
sorl-thumbnail                == 12.5.0
tesseract                     == 0.1.3
tqdm                          == 4.11.2
uritemplate                   == 3.0.0
urllib3                       == 1.24.1
django-crontab                == 0.7.1
